user/requirements.txt
Flask>=3.0.0
Pillow>=9.0.0
pytesseract>=0.3.10
